+++ source/Views/TextureView.cpp
@@ -21,6 +21,6 @@ void TextureView::draw_content()
 
     ImTextureID textureId = (void*)handle.ptr;
     
-    auto size = ImVec2(texture->Width, texture->Height);
+    auto size = ImVec2((float)texture->Width, (float)texture->Height);
     ImGui::Image(textureId, size, ImVec2(0, 1), ImVec2(1, 0));
 }
